Move RDS helpers to helpers.py + add 'make doctests'
