+++ aws/rds/test_rds_db_instance_not_publicly_accessible_by_vpc_sg.py
@@ -1,25 +1,11 @@
-
 import pytest
 
 from aws.rds.resources import (
     rds_db_instances,
     rds_db_instances_vpc_security_groups,
 )
-
-
-def does_vpc_security_group_grant_public_access(sg):
-    """
-    Checks an RDS instance for a VPC security groups with ingress permission ipv4 range 0.0.0.0/0 or ipv6 range :::/0
-
-    >>> does_vpc_security_group_grant_public_access({"IpPermissions":
-[{"CIDRIP": "127.0.0.1/32", "Status": "authorized"}, {"CIDRIP": "0.0.0.0/0", "Status": "authorized"}]})
-    True
-    >>> does_vpc_security_group_grant_public_access({"IpPermissions":
-IPRanges": []})
-    False
-    """
-    return any(ipr['CidrIp'] == '::/0' for ipp in sg['IpPermissions'] for ipr in ipp['Ipv6Ranges']) or \
-      any(ipr['CidrIp'] == '0.0.0.0/0' for ipp in sg['IpPermissions'] for ipr in ipp['IpRanges'])
+from aws.rds.helpers import \
+    does_vpc_security_group_grant_public_access
 
 
 @pytest.mark.rds
